Rename and fix bugs
